+++ app/client/src/routes/frf2022.tsx
@@ -9,7 +9,7 @@ import icons from "uswds/img/sprite.svg";
 // ---
 import { serverUrl, messages } from "../config";
 import {
-  FormioCRFSubmission,
+  FormioFRFSubmission,
   getData,
   postData,
   useContentData,
@@ -23,6 +23,7 @@ import {
 import { Loading } from "components/loading";
 import { Message } from "components/message";
 import { MarkdownContent } from "components/markdownContent";
+import { useDialogActions } from "contexts/dialog";
 import { useNotificationsActions } from "contexts/notifications";
 import { useRebateYearState } from "contexts/rebateYear";
 
@@ -35,24 +36,23 @@ type ServerResponse =
   | {
       userAccess: true;
       formSchema: { url: string; json: object };
-      submission: FormioCRFSubmission;
+      submission: FormioFRFSubmission;
     };
 
 /** Custom hook to fetch Formio submission data */
-function useFormioSubmissionQueryAndMutation(rebateId: string | undefined) {
+function useFormioSubmissionQueryAndMutation(mongoId: string | undefined) {
   const queryClient = useQueryClient();
 
   useEffect(() => {
-    queryClient.resetQueries({ queryKey: ["formio/2022/crf-submission"] });
+    queryClient.resetQueries({ queryKey: ["formio/2022/frf-submission"] });
   }, [queryClient]);
 
-  const url = `${serverUrl}/api/formio/2022/crf-submission/${rebateId}`;
+  const url = `${serverUrl}/api/formio/2022/frf-submission/${mongoId}`;
 
   const query = useQuery({
-    queryKey: ["formio/2022/crf-submission", { id: rebateId }],
+    queryKey: ["formio/2022/frf-submission", { id: mongoId }],
     queryFn: () => {
       return getData<ServerResponse>(url).then((res) => {
-        const mongoId = res.submission?._id;
         const comboKey = res.submission?.data.bap_hidden_entity_combo_key;
 
         /**
@@ -65,11 +65,19 @@ function useFormioSubmissionQueryAndMutation(rebateId: string | undefined) {
          */
         Formio.Providers.providers.storage.s3 = function (formio: any) {
           const s3Formio = cloneDeep(formio);
-          s3Formio.formUrl = `${serverUrl}/api/formio/2022/s3/crf/${mongoId}/${comboKey}`;
+          s3Formio.formUrl = `${serverUrl}/api/formio/2022/s3/frf/${mongoId}/${comboKey}`;
           return s3(s3Formio);
         };
 
-        return Promise.resolve(res);
+        // remove `ncesDataSource` and `ncesDataLookup` fields
+        const data = { ...res.submission?.data };
+        if (data.hasOwnProperty("ncesDataSource")) delete data.ncesDataSource;
+        if (data.hasOwnProperty("ncesDataLookup")) delete data.ncesDataLookup;
+
+        return Promise.resolve({
+          ...res,
+          submission: { ...res.submission, data },
+        });
       });
     },
     refetchOnWindowFocus: false,
@@ -77,18 +85,15 @@ function useFormioSubmissionQueryAndMutation(rebateId: string | undefined) {
 
   const mutation = useMutation({
     mutationFn: (updatedSubmission: {
-      mongoId: string;
-      submission: {
-        data: { [field: string]: unknown };
-        metadata: { [field: string]: unknown };
-        state: "submitted" | "draft";
-      };
+      data: { [field: string]: unknown };
+      metadata: { [field: string]: unknown };
+      state: "submitted" | "draft";
     }) => {
-      return postData<FormioCRFSubmission>(url, updatedSubmission);
+      return postData<FormioFRFSubmission>(url, updatedSubmission);
     },
     onSuccess: (res) => {
       return queryClient.setQueryData<ServerResponse>(
-        ["close-out", { id: rebateId }],
+        ["application", { id: mongoId }],
         (prevData) => {
           return prevData?.submission
             ? { ...prevData, submission: res }
@@ -101,24 +106,26 @@ function useFormioSubmissionQueryAndMutation(rebateId: string | undefined) {
   return { query, mutation };
 }
 
-export function CRF() {
+export function FRF2022() {
   const { email } = useOutletContext<{ email: string }>();
   /* ensure user verification (JWT refresh) doesn't cause form to re-render */
   return useMemo(() => {
-    return <CloseOutRequestForm email={email} />;
+    return <FundingRequestForm email={email} />;
   }, [email]);
 }
 
-function CloseOutRequestForm(props: { email: string }) {
+function FundingRequestForm(props: { email: string }) {
   const { email } = props;
 
   const navigate = useNavigate();
-  const { id: rebateId } = useParams<"id">(); // CSB Rebate ID (6 digits)
+  const { id: mongoId } = useParams<"id">(); // MongoDB ObjectId string
 
   const content = useContentData();
   const configData = useConfigData();
   const bapSamData = useBapSamData();
+  const { displayDialog } = useDialogActions();
   const {
+    displayInfoNotification,
     displaySuccessNotification,
     displayErrorNotification,
     dismissNotification,
@@ -128,7 +135,7 @@ function CloseOutRequestForm(props: { email: string }) {
   const submissionsQueries = useSubmissionsQueries();
   const rebates = useRebates();
 
-  const { query, mutation } = useFormioSubmissionQueryAndMutation(rebateId);
+  const { query, mutation } = useFormioSubmissionQueryAndMutation(mongoId);
   const { userAccess, formSchema, submission } = query.data ?? {};
 
   /**
@@ -182,23 +189,139 @@ function CloseOutRequestForm(props: { email: string }) {
     return <Message type="error" text={messages.formSubmissionError} />;
   }
 
-  const rebate = rebates.find((r) => r.rebateId === rebateId);
+  const rebate = rebates.find((r) => r.frf.formio._id === mongoId);
 
-  const crfNeedsEdits = !rebate
+  const frfNeedsEdits = !rebate
     ? false
     : submissionNeedsEdits({
-        formio: rebate.crf.formio,
-        bap: rebate.crf.bap,
+        formio: rebate.frf.formio,
+        bap: rebate.frf.bap,
       });
 
-  const crfSubmissionPeriodOpen =
-    configData.submissionPeriodOpen[rebateYear].crf;
+  const frfNeedsEditsAndPRFExists = frfNeedsEdits && !!rebate?.prf.formio;
+
+  /**
+   * NOTE: If the FRF submission needs edits and there's a corresponding PRF
+   * submission, display a confirmation dialog prompting the user to delete the
+   * PRF submission, as it's data will no longer valid when the FRF submission's
+   * data is changed.
+   */
+  if (frfNeedsEditsAndPRFExists) {
+    displayDialog({
+      dismissable: true,
+      heading: "Submission Edits Requested",
+      description: (
+        <>
+          <p>
+            This Application form submission has been opened at the request of
+            the applicant to make edits, but before you can make edits, the
+            associated Payment Request form submission needs to be deleted. If
+            the request to make edits to your Application form submission was
+            made in error, contact the Clean School Bus Program helpline at{" "}
+            <a href="mailto:cleanschoolbus@epa.gov">cleanschoolbus@epa.gov</a>.
+          </p>
+
+          <p>
+            If you’d like to view the Payment Request form submission before
+            deletion, please close this dialog box, and you will be re-directed
+            to the associated Payment Request form.
+          </p>
+
+          <p>
+            To proceed with deleting the associated Payment Request form
+            submission, please select the{" "}
+            <strong>Delete Payment Request Form Submission</strong> button
+            below, and the Payment Request form submission will be deleted. The
+            Application form will then be open for editing.
+          </p>
+
+          <div className="usa-alert usa-alert--error" role="alert">
+            <div className="usa-alert__body">
+              <p className="usa-alert__text">
+                <strong>Please note:</strong> Once deleted, the Payment Request
+                form submission will be removed from your dashboard and cannot
+                be recovered.
+              </p>
+            </div>
+          </div>
+        </>
+      ),
+      confirmText: "Delete Payment Request Form Submission",
+      confirmedAction: () => {
+        const prf = rebate.prf.formio;
+
+        if (!prf) {
+          displayErrorNotification({
+            id: Date.now(),
+            body: (
+              <>
+                <p className="tw-text-sm tw-font-medium tw-text-gray-900">
+                  Error deleting Payment Request <em>{rebate.rebateId}</em>.
+                </p>
+                <p className="tw-mt-1 tw-text-sm tw-text-gray-500">
+                  Please notify the helpdesk that a problem exists preventing
+                  the deletion of Payment Request form submission{" "}
+                  <em>{rebate.rebateId}</em>.
+                </p>
+              </>
+            ),
+          });
+
+          // NOTE: logging rebate for helpdesk debugging purposes
+          console.log(rebate);
+          return;
+        }
+
+        displayInfoNotification({
+          id: Date.now(),
+          body: (
+            <p className="tw-text-sm tw-font-medium tw-text-gray-900">
+              Deleting Payment Request <em>{rebate.rebateId}</em>...
+            </p>
+          ),
+        });
+
+        const url = `${serverUrl}/api/formio/2022/delete-prf-submission`;
+
+        postData(url, {
+          mongoId: prf._id,
+          rebateId: prf.data.hidden_bap_rebate_id,
+          comboKey: prf.data.bap_hidden_entity_combo_key,
+        })
+          .then((res) => {
+            window.location.reload();
+          })
+          .catch((err) => {
+            displayErrorNotification({
+              id: Date.now(),
+              body: (
+                <>
+                  <p className="tw-text-sm tw-font-medium tw-text-gray-900">
+                    Error deleting Payment Request <em>{rebate.rebateId}</em>.
+                  </p>
+                  <p className="tw-mt-1 tw-text-sm tw-text-gray-500">
+                    Please reload the page to attempt the deletion again, or
+                    contact the helpdesk if the problem persists.
+                  </p>
+                </>
+              ),
+            });
+          });
+      },
+      dismissedAction: () => navigate(`/payment-request/${rebate.rebateId}`),
+    });
+
+    return null;
+  }
+
+  const frfSubmissionPeriodOpen =
+    configData.submissionPeriodOpen[rebateYear].frf;
 
   const formIsReadOnly =
-    (submission.state === "submitted" || !crfSubmissionPeriodOpen) &&
-    !crfNeedsEdits;
+    (submission.state === "submitted" || !frfSubmissionPeriodOpen) &&
+    !frfNeedsEdits;
 
-  /** matched SAM.gov entity for the Close Out submission */
+  /** matched SAM.gov entity for the Application submission */
   const entity = bapSamData.entities.find((entity) => {
     const { ENTITY_COMBO_KEY__c } = entity;
     return ENTITY_COMBO_KEY__c === submission.data.bap_hidden_entity_combo_key;
@@ -221,9 +344,9 @@ function CloseOutRequestForm(props: { email: string }) {
           className="margin-top-4"
           children={
             submission.state === "draft"
-              ? content.draftCRFIntro
+              ? content.draftFRFIntro
               : submission.state === "submitted"
-              ? content.submittedCRFIntro
+              ? content.submittedFRFIntro
               : ""
           }
         />
@@ -237,9 +360,22 @@ function CloseOutRequestForm(props: { email: string }) {
             </svg>
           </div>
           <div className="usa-icon-list__content">
-            <strong>Rebate ID:</strong> {rebateId}
+            <strong>Application ID:</strong> {submission._id}
           </div>
         </li>
+
+        {rebate?.frf.bap?.rebateId && (
+          <li className="usa-icon-list__item">
+            <div className="usa-icon-list__icon text-primary">
+              <svg className="usa-icon" aria-hidden="true" role="img">
+                <use href={`${icons}#local_offer`} />
+              </svg>
+            </div>
+            <div className="usa-icon-list__content">
+              <strong>Rebate ID:</strong> {rebate.frf.bap.rebateId}
+            </div>
+          </li>
+        )}
       </ul>
 
       <Dialog as="div" open={dataIsPosting.current} onClose={(ev) => {}}>
@@ -286,12 +422,13 @@ function CloseOutRequestForm(props: { email: string }) {
 
             const data = { ...onSubmitSubmission.data };
 
+            // remove `ncesDataSource` and `ncesDataLookup` fields
+            if (data.hasOwnProperty("ncesDataSource")) delete data.ncesDataSource; // prettier-ignore
+            if (data.hasOwnProperty("ncesDataLookup")) delete data.ncesDataLookup; // prettier-ignore
+
             const updatedSubmission = {
-              mongoId: submission._id,
-              submission: {
-                ...onSubmitSubmission,
-                data,
-              },
+              ...onSubmitSubmission,
+              data,
             };
 
             dismissNotification({ id: 0 });
@@ -312,7 +449,7 @@ function CloseOutRequestForm(props: { email: string }) {
                     <p className="tw-text-sm tw-font-medium tw-text-gray-900">
                       {onSubmitSubmission.state === "submitted" ? (
                         <>
-                          Close Out <em>{rebateId}</em> submitted successfully.
+                          Application <em>{mongoId}</em> submitted successfully.
                         </>
                       ) : (
                         <>Draft saved successfully.</>
@@ -339,7 +476,7 @@ function CloseOutRequestForm(props: { email: string }) {
                   body: (
                     <p className="tw-text-sm tw-font-medium tw-text-gray-900">
                       {onSubmitSubmission.state === "submitted" ? (
-                        <>Error submitting Close Out form.</>
+                        <>Error submitting Application form.</>
                       ) : (
                         <>Error saving draft.</>
                       )}
@@ -364,6 +501,10 @@ function CloseOutRequestForm(props: { email: string }) {
 
             const data = { ...onNextPageParam.submission.data };
 
+            // remove `ncesDataSource` and `ncesDataLookup` fields
+            if (data.hasOwnProperty("ncesDataSource")) delete data.ncesDataSource; // prettier-ignore
+            if (data.hasOwnProperty("ncesDataLookup")) delete data.ncesDataLookup; // prettier-ignore
+
             // "dirty check" – don't post an update if no changes have been made
             // to the form (ignoring current user fields)
             const currentData = { ...data };
@@ -377,12 +518,9 @@ function CloseOutRequestForm(props: { email: string }) {
             if (isEqual(currentData, submittedData)) return;
 
             const updatedSubmission = {
-              mongoId: submission._id,
-              submission: {
-                ...onNextPageParam.submission,
-                data,
-                state: "draft" as const,
-              },
+              ...onNextPageParam.submission,
+              data,
+              state: "draft" as const,
             };
 
             dismissNotification({ id: 0 });
